// packages/frontend/src/types.ts
export interface Contact {
    id: string;
    name: string;
    birth: string;
    hobby: string
}
export interface Employee {
    employeeID: string;
    lastName: string;
    firstName: string;
    title: string;
    titleOfCourtesy: string;
    birthDate: string;
    hireDate: string;
    address: string;
    city: string;
    region: string;
    postalCode: string;
    country: string;
    homePhone: string;
    extension: string;
    photo: string;
    notes: string;
    reportsTo: string;
    photoPath: string;
  }
  